DPLAN-17480: extract token management logic out of ExpirationTimestampRequestListener
